Moved DataGenerator sketch from core.axis to core.data
